Extract mapping logic to specific classes
