+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/aas/EnvironmentToAssetMapper.java
@@ -15,57 +15,51 @@
  */
 package de.fraunhofer.iosb.app.aas;
 
+import de.fraunhofer.iosb.app.aas.mapper.ConceptDescriptionMapper;
+import de.fraunhofer.iosb.app.aas.mapper.ShellToAssetMapper;
+import de.fraunhofer.iosb.app.aas.mapper.SubmodelMapper;
 import de.fraunhofer.iosb.app.model.aas.service.Service;
 import de.fraunhofer.iosb.app.pipeline.PipelineFailure;
 import de.fraunhofer.iosb.app.pipeline.PipelineResult;
 import de.fraunhofer.iosb.app.pipeline.PipelineStep;
-import de.fraunhofer.iosb.dataplane.aas.spi.AasDataAddress;
 import org.eclipse.digitaltwin.aas4j.v3.model.AssetAdministrationShell;
 import org.eclipse.digitaltwin.aas4j.v3.model.ConceptDescription;
 import org.eclipse.digitaltwin.aas4j.v3.model.Environment;
-import org.eclipse.digitaltwin.aas4j.v3.model.Identifiable;
-import org.eclipse.digitaltwin.aas4j.v3.model.KeyTypes;
-import org.eclipse.digitaltwin.aas4j.v3.model.Referable;
-import org.eclipse.digitaltwin.aas4j.v3.model.Reference;
-import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
-import org.eclipse.digitaltwin.aas4j.v3.model.SubmodelElement;
-import org.eclipse.digitaltwin.aas4j.v3.model.SubmodelElementCollection;
-import org.eclipse.digitaltwin.aas4j.v3.model.SubmodelElementList;
-import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultAdministrativeInformation;
-import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultEmbeddedDataSpecification;
-import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultKey;
-import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultReference;
 import org.eclipse.edc.connector.controlplane.asset.spi.domain.Asset;
-import org.jetbrains.annotations.NotNull;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.function.Supplier;
 
 import static de.fraunhofer.iosb.app.pipeline.util.PipelineUtils.extractContents;
 import static de.fraunhofer.iosb.app.pipeline.util.PipelineUtils.handleError;
 
 /**
- * Create a mapping from an AAS environment to EDC assets.
- * This is not a holistic transformation but rather maps some
- * key elements and creates appropriate data address and assetId.
+ * Create a mapping from an AAS environment to EDC assets. This is not a holistic transformation but rather maps key
+ * elements and creates appropriate data address and assetId.
+ * <p>
+ * Here, we control what information about the AAS elements is stored in the AssetIndex and in the extension.
  */
 public class EnvironmentToAssetMapper extends PipelineStep<Map<Service, Environment>, Collection<Service>> {
 
     private static final String CONCEPT_DESCRIPTIONS = "conceptDescriptions";
     private static final String SHELLS = "shells";
     private static final String SUBMODELS = "submodels";
+
     private final Supplier<Boolean> onlySubmodelsDecision;
+    private final ShellToAssetMapper shellMapper = new ShellToAssetMapper();
+    private final SubmodelMapper submodelMapper;
+    private final ConceptDescriptionMapper conceptDescriptionMapper = new ConceptDescriptionMapper();
+
 
     public EnvironmentToAssetMapper(Supplier<Boolean> onlySubmodelsDecision) {
         this.onlySubmodelsDecision = onlySubmodelsDecision;
+        submodelMapper = new SubmodelMapper(onlySubmodelsDecision);
     }
 
     /**
@@ -104,7 +98,7 @@ public class EnvironmentToAssetMapper extends PipelineStep<Map<Service, Environm
         try {
             final String submodelsUrl = service.getSubmodelsUrl().toString();
             assetBuilder.property(SUBMODELS, environment.getSubmodels().stream()
-                    .map(submodel -> mapSubmodelToAsset(submodel, submodelsUrl))
+                    .map(submodel -> submodelMapper.map(submodel, submodelsUrl))
                     .toList());
         } catch (MalformedURLException e) {
             return PipelineResult.failure(PipelineFailure.warning(List.of("Could not build access url for %s".formatted(service.getAccessUrl()))));
@@ -124,142 +118,16 @@ public class EnvironmentToAssetMapper extends PipelineStep<Map<Service, Environm
                     assetBuilder
                             .property(SHELLS,
                                     environment.getAssetAdministrationShells().stream()
-                                            .map((AssetAdministrationShell shell) -> mapShellToAsset(shell, shellsUrl))
+                                            .map((AssetAdministrationShell shell) -> shellMapper.map(shell, shellsUrl))
                                             .toList())
                             .property(CONCEPT_DESCRIPTIONS,
                                     environment.getConceptDescriptions().stream()
-                                            .map((ConceptDescription conceptDescription) -> mapConceptDescriptionToAsset(conceptDescription, conceptDescriptionsUrl))
+                                            .map((ConceptDescription conceptDescription) -> conceptDescriptionMapper.map(conceptDescription, conceptDescriptionsUrl))
                                             .toList())
                             .build()));
         } catch (MalformedURLException e) {
             return PipelineResult.recoverableFailure(service.with(assetBuilder.build()),
                     PipelineFailure.warning(List.of("Could not build access url for %s".formatted(service.getAccessUrl()))));
         }
-    }
-
-    private <R extends Referable> Asset.Builder mapReferableToAssetBuilder(R referable) {
-        return Asset.Builder.newInstance()
-                .properties(Map.of(
-                        "idShort", Optional.ofNullable(referable.getIdShort()).orElse(""),
-                        "name", referable.getDisplayName(),
-                        "description", referable.getDescription()));
-    }
-
-    private <I extends Identifiable> Asset.Builder mapIdentifiableToAssetBuilder(I identifiable) {
-        var admin = Optional.ofNullable(identifiable.getAdministration())
-                .orElse(new DefaultAdministrativeInformation.Builder().build());
-        var version = null != admin.getVersion() && null != admin.getRevision() ?
-                String.valueOf(admin.getVersion()).concat(":").concat(String.valueOf(admin.getRevision())) : null;
-
-        return mapReferableToAssetBuilder(identifiable)
-                .version(version)
-                .contentType("application/json")
-                .properties(Map.of(
-                        "id", identifiable.getId(),
-                        "embeddedDataSpecifications", admin.getEmbeddedDataSpecifications()));
-    }
-
-    /* May contain traces of recursion */
-    private <E extends SubmodelElement> Asset mapSubmodelElementToAsset(Reference parentReference, E submodelElement,
-                                                                        String accessUrl) {
-        var reference = createReference(submodelElement.getIdShort(), parentReference);
-
-        var children = getContainerElements(submodelElement).stream()
-                .map(elem -> mapSubmodelElementToAsset(reference, elem, accessUrl))
-                .toList();
-
-        var dataAddress = AasDataAddress.Builder.newInstance()
-                .baseUrl(accessUrl)
-                .referenceChain(reference)
-                .build();
-
-        return mapReferableToAssetBuilder(submodelElement)
-                .id(getId(accessUrl, dataAddress))
-                .contentType("application/json")
-                .properties(Map.of(
-                        "embeddedDataSpecifications",
-                        Optional.ofNullable(submodelElement.getEmbeddedDataSpecifications())
-                                .orElse(List.of(new DefaultEmbeddedDataSpecification())),
-                        "semanticId",
-                        Optional.ofNullable(submodelElement.getSemanticId())
-                                .orElse(new DefaultReference()),
-                        "value", children))
-                .dataAddress(dataAddress)
-                .build();
-    }
-
-    private static @NotNull String getId(String accessUrl, AasDataAddress dataAddress) {
-        return String.valueOf("%s:%s".formatted(accessUrl, dataAddress.getPath()).hashCode());
-    }
-
-    private <T extends SubmodelElement> Collection<SubmodelElement> getContainerElements(T submodelElement) {
-        if (submodelElement instanceof SubmodelElementCollection) {
-            return ((SubmodelElementCollection) submodelElement).getValue();
-        } else if (submodelElement instanceof SubmodelElementList) {
-            return ((SubmodelElementList) submodelElement).getValue();
-        } else { // Can not have any child elements...
-            return Collections.emptyList();
-        }
-    }
-
-    private Asset mapShellToAsset(AssetAdministrationShell shell, String accessUrl) {
-        var dataAddress = AasDataAddress.Builder.newInstance()
-                .baseUrl(accessUrl)
-                .referenceChain(createReference(KeyTypes.ASSET_ADMINISTRATION_SHELL, shell.getId()))
-                .build();
-
-        return mapIdentifiableToAssetBuilder(shell)
-                .id(getId(accessUrl, dataAddress))
-                .dataAddress(dataAddress)
-                .build();
-    }
-
-    private Asset mapConceptDescriptionToAsset(ConceptDescription conceptDescription, String accessUrl) {
-        var dataAddress = AasDataAddress.Builder.newInstance()
-                .baseUrl(accessUrl)
-                .referenceChain(createReference(KeyTypes.CONCEPT_DESCRIPTION, conceptDescription.getId()))
-                .build();
-
-        return mapIdentifiableToAssetBuilder(conceptDescription)
-                .id(getId(accessUrl, dataAddress))
-                .dataAddress(dataAddress)
-                .build();
-    }
-
-    private Asset mapSubmodelToAsset(Submodel submodel, String accessUrl) {
-        var reference = createReference(KeyTypes.SUBMODEL, submodel.getId());
-        List<Asset> children = new ArrayList<>();
-        if (!onlySubmodelsDecision.get()) {
-            children = submodel.getSubmodelElements().stream()
-                    .map(elem -> mapSubmodelElementToAsset(reference, elem, accessUrl))
-                    .toList();
-        }
-
-        var dataAddress = AasDataAddress.Builder.newInstance()
-                .baseUrl(accessUrl).referenceChain(reference)
-                .build();
-
-        return mapIdentifiableToAssetBuilder(submodel)
-                .id(getId(accessUrl, dataAddress))
-                .properties(Map.of(
-                        "semanticId",
-                        Optional.ofNullable(submodel.getSemanticId())
-                                .orElse(new DefaultReference.Builder().build()),
-                        "submodelElements", children))
-                .dataAddress(dataAddress)
-                .build();
-    }
-
-    private Reference createReference(KeyTypes type, String value) {
-        return new DefaultReference.Builder()
-                .keys(new DefaultKey.Builder().type(type).value(value).build())
-                .build();
-    }
-
-    private Reference createReference(String value, Reference parent) {
-        return new DefaultReference.Builder()
-                .keys(new ArrayList<>(parent.getKeys()))
-                .keys(new DefaultKey.Builder().type(KeyTypes.SUBMODEL_ELEMENT).value(value).build())
-                .build();
     }
 }
